Q4 complete
Competed plotting a conditional time series plot of mean steps by
interval for weekdays and weekends.
This took more stages than I initially hoped. Was unable to create the
factors directly from the day of the week. Firstly I specified the
appropriate string, then converted all the strings to factors
afterwards,
Following this, the mean step counts for each interval were calculated,
but separated out by type of day of the week in addition to interval,
unlike in Q2.

Plotting was implemented using ggplot2 package, similarly to Q2 but this
time specifying a faceting on type of day of the week.

This completes assignment 1.